feat: support history forget timeout in agent runner
